diarization_service/requirements.txt
flask==2.3.2
torch
pyannote.audio
speechbrain==1.0.2
pydub==0.25.1
librosa==0.9.2
numpy==1.24.3
requests==2.31.0

